Clientside JS (#94)
* Remove semicolons

* Remove configureEnginePlugin from exports

* Split up application.js

* Rename to preview-close-link

* Remove file-upload from webpack

* Export an init function from js files

* Remove file upload from package.json

* Remove inline script from file-upload

* Update package-lock.json

* Add shared client js file
